+++ src/Smuggler2/Options.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DeriveAnyClass #-}
 -- |
 -- Description: handling of command line options
 module Smuggler2.Options
@@ -8,10 +9,12 @@ module Smuggler2.Options
   )
 where
 
-import Data.Char (toLower)
-import Data.List (foldl', isPrefixOf)
-import Data.List.Split (splitOn)
-import Plugins (CommandLineOption)
+import Data.Char ( isSpace, toLower )
+import Data.List ( foldl' )
+import Data.List.Split ( splitOn )
+import GHC ( mkModuleName, ModuleName )
+import Outputable ( Outputable )
+import Plugins ( CommandLineOption )
 
 -- | Ways of performing import processing
 data ImportAction = NoImportProcessing | PreserveInstanceImports | MinimiseImports
@@ -26,14 +29,15 @@ data Options = Options
   { importAction :: ImportAction,
     exportAction :: ExportAction,
     newExtension :: Maybe String,
-    leaveOpenImports :: [String]
+    leaveOpenImports :: [ModuleName],
+    makeOpenImports :: [ModuleName]
   }
-  deriving (Show)
+  deriving (Outputable)
 
 -- | The default is to retain instance-only imports (eg, Data.List () )
 -- and add explict exports only if they are not already present
 defaultOptions :: Options
-defaultOptions = Options PreserveInstanceImports AddExplicitExports Nothing []
+defaultOptions = Options PreserveInstanceImports AddExplicitExports Nothing [] []
 
 -- | Simple command line option parser.  Last occurrence wins.
 parseCommandLineOptions :: [CommandLineOption] -> Options
@@ -47,11 +51,20 @@ parseCommandLineOptions = foldl' parseCommandLineOption defaultOptions
       "noexportprocessing" -> opts {exportAction = NoExportProcessing}
       "addexplicitexports" -> opts {exportAction = AddExplicitExports}
       "replaceexports" -> opts {exportAction = ReplaceExports}
-      o ->
-        if "leaveopenimports" `isPrefixOf` o
-          then
-            opts
-              { leaveOpenImports =
-                   splitOn "," (drop (length "leaveopenimports:") clo)
-              }
-          else opts {newExtension = Just clo}
+      _
+        | Just modulenames <- stripPrefixCI "leaveopenimports:" clo ->
+          opts {leaveOpenImports = parseModuleNames modulenames}
+        | Just modulenames <- stripPrefixCI "makeopenimports:" clo ->
+          opts {makeOpenImports = parseModuleNames modulenames}
+        | otherwise -> opts {newExtension = Just clo}
+
+-- | split on comma.  Not v robust (
+parseModuleNames :: String -> [ModuleName]
+parseModuleNames arg = mkModuleName <$> splitOn "," (filter (not . isSpace) arg)
+
+-- | case-insensitive version of @Data.List.stripPrefix@
+stripPrefixCI :: String -> String -> Maybe String
+stripPrefixCI [] ys = Just ys
+stripPrefixCI (x : xs) (y : ys)
+  | toLower x == toLower y = stripPrefixCI xs ys
+stripPrefixCI _ _ = Nothing