+++ tmp/Operator.hs
@@ -1,0 +1,8 @@
+module Operator where
+
+infixl 5 $+$
+
+($+$) :: [a] -> [a] -> [a]
+[] $+$ d = d
+d $+$ [] = d
+d $+$ d' = d ++ d'